# internal/api/openapi.yaml
openapi: "3.1.0"
info:
  title: sqi-server API
  description: |
    REST API for the sqi distributed task management server.

    All error responses use the RFC 7807 problem-details format
    (`application/problem+json`).  Clients should branch on `Content-Type`
    rather than status code alone to detect error bodies.
  version: "0.1.0"
  license:
    name: AGPL-3.0-or-later
    url: https://www.gnu.org/licenses/agpl-3.0.html
servers:
  - url: /api/v1
    description: Current server
...
components:
  parameters:
    id:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Resource UUID
...
  schemas:
...
    ProblemDetail:
      type: object
      required: [type, title, status, detail]
      properties:
        type:
          type: string
          example: "about:blank"
        title:
          type: string
          example: Not Found
        status:
          type: integer
          example: 404
        detail:
          type: string
          example: job not found
        instance:
          type: string
          description: Request-ID that identifies this specific occurrence.
          example: a1b2c3d4e5f60708

    # ── Jobs ─────────────────────────────────────────────────────────────────
...
    StorageLocation:
      type: object
      required: [id, name, type, roots, created_at, updated_at]
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
          enum: [filesystem, s3]
        description:
          type: string
        roots:
          type: object
          additionalProperties:
            type: string
          description: |
            Map of compute-location name → absolute root path for that location.
            Example: `{"on-prem": "/mnt/farm", "aws-us-east-1": "s3://bucket/root"}`
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateStorageLocationRequest:
      type: object
      required: [name, type]
      properties:
        name:
          type: string
        type:
          type: string
          enum: [filesystem, s3]
        description:
          type: string
        roots:
          type: object
          additionalProperties:
            type: string
    UpdateStorageLocationRequest:
      type: object
      required: [name, type]
      properties:
        name:
          type: string
        type:
          type: string
          enum: [filesystem, s3]
        description:
          type: string
        roots:
          type: object
          additionalProperties:
            type: string

    # ── License pools ─────────────────────────────────────────────────────────
...
  responses:
    BadRequest:
      description: Bad request — missing or invalid parameters.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
    NotFound:
      description: Resource not found.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
    Conflict:
      description: Conflict — unique constraint violated or invalid state transition.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
...
    InternalServerError:
      description: Internal server error.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"

# ── Paths ─────────────────────────────────────────────────────────────────────
paths:
...
  /storage-locations:
    post:
      operationId: createStorageLocation
      summary: Create a storage location
      tags: [storage-locations]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStorageLocationRequest"
      responses:
        "201":
          description: Storage location created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StorageLocation"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalServerError"

    get:
      operationId: listStorageLocations
      summary: List all storage locations
      tags: [storage-locations]
      responses:
        "200":
          description: Array of storage locations.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StorageLocation"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /storage-locations/{id}:
    parameters:
      - $ref: "#/components/parameters/id"

    get:
      operationId: getStorageLocation
      summary: Get a storage location
      tags: [storage-locations]
      responses:
        "200":
          description: Storage location.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StorageLocation"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

    put:
      operationId: updateStorageLocation
      summary: Update a storage location
      tags: [storage-locations]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStorageLocationRequest"
      responses:
        "200":
          description: Updated storage location.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StorageLocation"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalServerError"

    delete:
      operationId: deleteStorageLocation
      summary: Delete a storage location
      tags: [storage-locations]
      responses:
        "204":
          description: Storage location deleted.
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # ── License pools ──────────────────────────────────────────────────────────
...